requirements.txt
click>=8.1
pyyaml>=6.0
pytest>=8.0
requests>=2.32
flask>=3.0
playwright>=1.43
PyPDF2>=3.0
